refactor: Organize project structure
This commit refactors the project structure to be more organized and scalable.

- Created a more modular structure with `app`, `shared`, and `components` directories.
- Moved pages, styles, and the main `App.tsx` into the `app` directory.
- Moved reusable hooks, utils, assets, and libraries into the `shared` directory.
- Separated common components from `shadcn/ui` components.
- Updated import paths across the application to reflect the new structure.
- Added new path aliases to `vite.config.ts` and `tsconfig.json` for easier imports.
- Removed old, unused directories.
