+++ src/components/common/AttractionCard.tsx
