+++ src/components/common/SearchSection.tsx
