+++ src/shared/assets/floating-market.jpg
