+++ src/shared/assets/hero-beach.jpg
