+++ src/shared/hooks/use-mobile.tsx
